requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-nmap==0.7.1
psutil==5.9.8
ping3==4.0.4
requests==2.31.0
mss==9.0.1
dicttoxml==1.7.16
ftplib
